swarms-agents/requirements.txt
annotated-types==0.7.0
anthropic==0.87.0
anyio==4.13.0
attrs==26.1.0
babel==2.18.0
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
coloredlogs==15.0.1
cryptography==46.0.6
csvw==3.7.0
ddgs==9.12.0
deprecation==2.1.0
distro==1.9.0
dlinfo==2.0.0
docstring_parser==0.17.0
duckduckgo_search==8.1.1
espeakng-loader==0.2.4
exceptiongroup==1.3.1
flatbuffers==25.12.19
fsspec==2026.3.0
google-api-core==2.30.1
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.1
googleapis-common-protos==1.73.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.11
isodate==0.7.2
jiter==0.13.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kokoro-onnx==0.5.0
language-tags==1.2.0
lxml==6.0.2
markdown-it-py==4.0.0
mdurl==0.1.2
mmh3==5.2.1
mpmath==1.3.0
multidict==6.7.1
numpy==2.2.6
onnxruntime==1.23.2
openai==2.30.0
packaging==26.0
phonemizer-fork==3.3.2
postgrest==2.28.3
primp==1.2.1
propcache==0.4.1
proto-plus==1.27.2
protobuf==6.33.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
pyiceberg==0.11.1
PyJWT==2.12.1
pyparsing==3.3.2
pyroaring==1.0.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
rdflib==7.6.0
realtime==2.28.3
referencing==0.37.0
regex==2026.3.32
requests==2.33.1
rfc3986==1.5.0
rich==14.3.3
rpds-py==0.30.0
scipy==1.15.3
segments==2.4.0
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
storage3==2.28.3
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.3
supabase-auth==2.28.3
supabase-functions==2.28.3
sympy==1.14.0
tenacity==9.1.4
termcolor==3.3.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
websockets==15.0.1
yarl==1.23.0
zstandard==0.25.0
